PACKAGE: restructure package using /lib
