+++ lib/rdfstore.js
@@ -1,0 +1,5 @@
+
+'use strict';
+
+const QuadStore = require('./quadstore');
+
